misc refactoring
