+++ StockApp/Classes/FinnHubCompanyResult.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
+﻿using StockApp.Interfaces;
 
 namespace StockApp.Classes
 {
