+++ StockApp/Interfaces/IFinnHubCompanyResult.cs
@@ -1,9 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
+﻿
 
-namespace StockApp.Classes
+namespace StockApp.Interfaces
 {
 	interface IFinnHubCompanyResult
 	{
