+++ StockApp/StockApp.csproj
@@ -90,7 +90,6 @@
     <Compile Include="Interfaces\IFinnHubCompanyResult.cs" />
     <Compile Include="Interfaces\IFinnHubQuoteResult.cs" />
     <Compile Include="Classes\Quote.cs" />
-    <Compile Include="Classes\SearchSymbol.cs" />
     <Compile Include="Classes\ServiceRequest.cs" />
     <Compile Include="Default.aspx.cs">
       <DependentUpon>Default.aspx</DependentUpon>
@@ -115,6 +114,7 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="App_Data\" />
+    <Folder Include="Enumerations\" />
     <Folder Include="Models\" />
   </ItemGroup>
   <PropertyGroup>
