+++ StockApp/bin/StockApp.dll.config
@@ -15,27 +15,24 @@
       <compiler language="vb;vbs;visualbasic;vbscript" extension=".vb" type="Microsoft.CodeDom.Providers.DotNetCompilerPlatform.VBCodeProvider, Microsoft.CodeDom.Providers.DotNetCompilerPlatform, Version=2.0.1.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35" warningLevel="4" compilerOptions="/langversion:default /nowarn:41008 /define:_MYTYPE=\&quot;Web\&quot; /optionInfer+" />
     </compilers>
   </system.codedom>
-
   <runtime>
-
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-
       <dependentAssembly>
-
         <assemblyIdentity name="Newtonsoft.Json" publicKeyToken="30ad4fe6b2a6aeed" culture="neutral" />
-
         <bindingRedirect oldVersion="0.0.0.0-12.0.0.0" newVersion="12.0.0.0" />
-
       </dependentAssembly>
-
     </assemblyBinding>
-
   </runtime>
-<system.webServer>
+  <system.webServer>
     <handlers>
       <remove name="ExtensionlessUrlHandler-Integrated-4.0" />
       <remove name="OPTIONSVerbHandler" />
       <remove name="TRACEVerbHandler" />
       <add name="ExtensionlessUrlHandler-Integrated-4.0" path="*." verb="*" type="System.Web.Handlers.TransferRequestHandler" preCondition="integratedMode,runtimeVersionv4.0" />
     </handlers>
-  </system.webServer></configuration>
+  </system.webServer>
+  <appSettings>
+    <add key="FinnHubBaseUrl" value="https://finnhub.io/api/v1/" />
+    <add key="FinnHubApiToken" value="bu7044748v6rghl7nkng" />
+  </appSettings>
+</configuration>
